requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
opencv-python>=4.5.0
numpy>=1.19.0
scipy>=1.7.0
onnxruntime==1.17.1
onnx==1.16.0
pillow==10.3.0
tqdm>=4.50.0
PyYAML>=5.3.1
filterpy>=1.4.5
torch>=1.8.0
torchvision>=0.9.0
ultralytics>=8.0.0
deep-sort-realtime>=1.3.0
opencv-contrib-python>=4.5.0
python-multipart==0.0.9
pytest==8.2.0
